useContext for cart
