+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
 import { useTheme } from '@/context/ThemeContext.tsx';
 
@@ -11,31 +11,17 @@ import styles from './App.module.css';
 
 function App() {
     const [currentComponent, setCurrentComponent] = useState<'About' | 'ProductsList'>('About');
-    const [cartCount, setCartCount] = useState<number>(0);
     const { theme } = useTheme();
 
     const handleChangePage = (component: 'About' | 'ProductsList') => {
         setCurrentComponent(component);
     };
 
-    useEffect(() => {
-        const storedCartCount = localStorage.getItem('cartCount');
-        if (storedCartCount) {
-            setCartCount(Number.parseInt(storedCartCount, 10));
-        }
-    }, []);
-
-    const addToCart = () => {
-        const newCartCount = cartCount + 1;
-        setCartCount(newCartCount);
-        localStorage.setItem('cartCount', newCartCount.toString());
-    };
-
     return (
         <>
             <div className={`${theme === 'dark' ? styles.darkTheme : styles.lightTheme}`}>
-                <HeaderComponent handleChangePage={handleChangePage} cartCount={cartCount} />
-                {currentComponent === 'About' ? <AboutPage /> : <ProductsList addToCart={addToCart} />}
+                <HeaderComponent handleChangePage={handleChangePage} />
+                {currentComponent === 'About' ? <AboutPage /> : <ProductsList />}
                 <FooterComponent />
             </div>
         </>
