+++ src/main.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
+import { CartProvider } from '@/context/CartContext.tsx';
 import { ThemeProvider } from '@/context/ThemeContext.tsx';
 
 import App from './App.tsx';
@@ -10,7 +11,9 @@ import './index.css';
 ReactDOM.createRoot(document.querySelector('#root')!).render(
     <React.StrictMode>
         <ThemeProvider>
-            <App />
+            <CartProvider>
+                <App />
+            </CartProvider>
         </ThemeProvider>
     </React.StrictMode>,
 );
